flask_server/requirements.txt
appengine-python-standard==0.2.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
docopt==0.6.2
Flask==2.0.2
frozendict==2.0.6
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mock==4.0.3
Pillow==8.4.0
pipreqs==0.4.11
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.15.100
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
zipp==3.6.0
